feat(document colors): ANSI escape colors

+++ lsp/textdocument_document_color.go
@@ -20,8 +20,8 @@ type ColorInformation struct {
 }
 
 type Color struct {
-	Red   float64 `json:"red"`
-	Green float64 `json:"green"`
-	Blue  float64 `json:"blue"`
-	Alpha float64 `json:"alpha"`
+	Red   float32 `json:"red"`
+	Green float32 `json:"green"`
+	Blue  float32 `json:"blue"`
+	Alpha float32 `json:"alpha"`
 }
